Refactor for PSR4 compliance
