+++ Tests/DependencyInjection/Compiler/FakeCompilerPass.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace Overblog\GraphQLBundle\Tests\DependencyInjection\Compiler;
+
+use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Reference;
+
+/**
+ * Class FakeCompilerPass
+ * @package DependencyInjection\Compiler
+ */
+class FakeCompilerPass implements CompilerPassInterface
+{
+    public function process(ContainerBuilder $container)
+    {
+        $container
+            ->getDefinition('test_resolver')
+            ->addArgument(new Reference('injected_service'))
+        ;
+    }
+}
