+++ Tests/DependencyInjection/Compiler/FakeInjectedService.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Overblog\GraphQLBundle\Tests\DependencyInjection\Compiler;
+
+/**
+ * Class FakeInjectedService
+ */
+class FakeInjectedService
+{
+    public function doSomething()
+    {
+        return true;
+    }
+}
