+++ Tests/DependencyInjection/Compiler/ResolverTaggedServiceMappingPassTest.php
@@ -3,12 +3,8 @@
 namespace Overblog\GraphQLBundle\Tests\DependencyInjection\Compiler;
 
 use Overblog\GraphQLBundle\DependencyInjection\Compiler\ResolverTaggedServiceMappingPass;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareTrait;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Definition;
-use Symfony\Component\DependencyInjection\Reference;
 
 /**
  * Class ResolverTaggedServiceMappingPassTest
@@ -34,13 +30,13 @@ class ResolverTaggedServiceMappingPassTest extends \PHPUnit_Framework_TestCase
             "Overblog\\GraphQLBundle\\Resolver\\ResolverResolver"
         );
 
-        $fakeResolver = new Definition('Overblog\GraphQLBundle\Tests\DependencyInjection\Compiler\FakeResolverService');
-        $fakeResolver
+        $testResolver = new Definition('Overblog\GraphQLBundle\Tests\DependencyInjection\Compiler\ResolverTestService');
+        $testResolver
             ->addTag( 'overblog_graphql.resolver', [
-                'alias' => 'fake_resolver', 'method' => 'doSomethingWithContainer'
+                'alias' => 'test_resolver', 'method' => 'doSomethingWithContainer'
             ]);
 
-        $container->setDefinition('fake_resolver', $fakeResolver);
+        $container->setDefinition('test_resolver', $testResolver);
 
         $this->container = $container;
     }
@@ -48,7 +44,6 @@ class ResolverTaggedServiceMappingPassTest extends \PHPUnit_Framework_TestCase
     private function addCompilerPassesAndCompile()
     {
             $this->container->addCompilerPass(new ResolverTaggedServiceMappingPass());
-            // Manipulate container after ResolverTaggedServiceMappingPass
             $this->container->addCompilerPass(new FakeCompilerPass());
             $this->container->compile();
     }
@@ -57,40 +52,6 @@ class ResolverTaggedServiceMappingPassTest extends \PHPUnit_Framework_TestCase
     {
         $this->addCompilerPassesAndCompile();
 
-        $this->assertTrue($this->container->has('fake_resolver'));
-    }
-}
-
-class FakeCompilerPass implements CompilerPassInterface
-{
-    public function process(ContainerBuilder $container)
-    {
-        $container
-            ->getDefinition('fake_resolver')
-            ->addArgument(new Reference('injected_service'))
-        ;
-    }
-}
-
-// Instantiate ContainerAwareInterface to trigger the problem in ResolverTaggedServiceMappingPass
-class FakeResolverService implements ContainerAwareInterface
-{
-    use ContainerAwareTrait;
-
-    public function __construct($service)
-    {
-    }
-
-    public function doSomethingWithContainer()
-    {
-        return $this->container->get('injected_service')->doSomething();
-    }
-}
-
-class FakeInjectedService
-{
-    public function doSomething()
-    {
-        return true;
+        $this->assertTrue($this->container->has('test_resolver'));
     }
 }
